feat: some improvement to Lua side
Related to #7

1. The `plugin/` directory is traditionally used to "bootstrap" a Neo(Vim)
plugin, but it is left empty.

My solution is to move some logic from `setup()` function to
`plugin/cursortab.lua`

2. There should be some type annotations in the Lua code, so that Lua
   language server can help you check types

3. Regarding configuration:

I think in the first release, we should limit what can be configured.
Only add more configuration options when users ask for it, because
configuration is also a maintainability problem.

For example, `command_name` shouldn't be configurable, as it will
complicate documentation, also most plugins don't allow Ex commands to
be customized.

Instead, keymaps should be configurable.